+++ src/gridjob_sample_posterior.sge
@@ -1,13 +1,13 @@
 #!/bin/bash
 #$ -N RKI_MCMC
 #$ -t 1-9
-#$ -l mem=20G
+#$ -l mem=14G
 #$ -m n
 #$ -q ni.q
 #$ -l h='!(ramsauer.ikw.uni-osnabrueck.de|vector.cv.uni-osnabrueck.de|righty.ni.uni-osnabrueck.de)'
 #$ -wd /net/store/ni/projects/BSTIM/src/
-#$ -e /net/store/ni/projects/BSTIM/logs/e_$TASK_ID.txt
-#$ -o /net/store/ni/projects/BSTIM/logs/o_$TASK_ID.txt
+#$ -e /net/store/ni/projects/BSTIM/logs_backup/e_$TASK_ID.txt
+#$ -o /net/store/ni/projects/BSTIM/logs_backup/o_$TASK_ID.txt
 #$ -pe default 4
 
 echo "Running job ${JOB_ID}, task ${SGE_TASK_ID} on `hostname`."
@@ -15,4 +15,5 @@ echo "Running job ${JOB_ID}, task ${SGE_TASK_ID} on `hostname`."
 mkdir "/tmp/${JOB_ID}_${SGE_TASK_ID}"
 
 source /net/store/users/jleugeri/virt/bin/activate
-THEANO_FLAGS="base_compiledir=/tmp/${JOB_ID}_${SGE_TASK_ID}/,floatX=float32,device=cpu,openmp=True,mode=FAST_RUN,warn_float64=warn" OMP_NUM_THREADS=4 python3 sample_posterior.py
+#THEANO_FLAGS="base_compiledir=/tmp/${JOB_ID}_${SGE_TASK_ID}/,floatX=float32,device=cpu,openmp=True,mode=FAST_RUN,warn_float64=warn" OMP_NUM_THREADS=4 python3 sample_posterior.py
+THEANO_FLAGS="base_compiledir=/tmp/${JOB_ID}_${SGE_TASK_ID}/,floatX=float32,device=cpu,mode=FAST_RUN,warn_float64=warn" python3 sample_posterior.py
